refactor: 패키지 구조 변경
- 횡단 관심사인 프레임워크 코드와 컨트롤러 코드 분리
